database edits
made new functions and changed how the database works for the different types of tags. Instead of having each tag in their own column, they're stored as a single string. Also made a counter for works written and bookmarked instead of scraping each title and link because those were not being used.